requirements.txt
setuptools~=46.1.3
pytest~=5.4.1
gitPython
